requirements.txt
requests==2.31.0
pandas==2.1.4
mysql-connector-python==8.2.0
vaderSentiment==3.3.2
yake==0.4.8
streamlit==1.31.0
plotly==5.18.0
wordcloud==1.9.3
matplotlib==3.8.2
apscheduler==3.10.4
python-dotenv==1.0.0
Pillow==10.2.0